packages/ingest-lib/pyproject.toml
[project]
name = "ingest-lib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adam Baxter", email = "ambaxter@github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "markdownify>=1.2.2",
    "milvus-lib",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[tool.uv.sources]
milvus-lib = { workspace = true }
